+++ src/cpp/session/include/session/SessionOptions.hpp
@@ -422,13 +422,6 @@ public:
       return core::string_utils::LineEndingPosix;
    }
 
-   // The line ending we persist to disk with. This could potentially
-   // be a per-user or even per-file option.
-   core::string_utils::LineEnding sourcePersistLineEnding() const
-   {
-      return core::string_utils::LineEndingNative;
-   }
-
    std::string monitorSharedSecret() const
    {
       return monitorSharedSecret_.c_str();