+++ src/gwt/src/org/rstudio/studio/client/workbench/prefs/views/EditingPreferencesPane.java
@@ -28,28 +28,38 @@ import com.google.gwt.user.client.ui.Label;
 import com.google.gwt.user.client.ui.VerticalPanel;
 import com.google.inject.Inject;
 
+import org.rstudio.core.client.StringUtil;
 import org.rstudio.core.client.prefs.PreferencesDialogBaseResources;
 import org.rstudio.core.client.theme.DialogTabLayoutPanel;
 import org.rstudio.core.client.widget.HelpButton;
 import org.rstudio.core.client.widget.ModifyKeyboardShortcutsWidget;
 import org.rstudio.core.client.widget.NumericValueWidget;
+import org.rstudio.core.client.widget.OperationWithInput;
 import org.rstudio.core.client.widget.SelectWidget;
 import org.rstudio.core.client.widget.SmallButton;
+import org.rstudio.core.client.widget.TextBoxWithButton;
 import org.rstudio.studio.client.common.DiagnosticsHelpLink;
 import org.rstudio.studio.client.common.HelpLink;
+import org.rstudio.studio.client.common.SimpleRequestCallback;
 import org.rstudio.studio.client.rmarkdown.RmdOutput;
+import org.rstudio.studio.client.workbench.prefs.model.EditingPrefs;
 import org.rstudio.studio.client.workbench.prefs.model.RPrefs;
 import org.rstudio.studio.client.workbench.prefs.model.UIPrefs;
 import org.rstudio.studio.client.workbench.prefs.model.UIPrefsAccessor;
 import org.rstudio.studio.client.workbench.snippets.ui.EditSnippetsDialog;
+import org.rstudio.studio.client.workbench.views.source.editors.text.IconvListResult;
+import org.rstudio.studio.client.workbench.views.source.editors.text.ui.ChooseEncodingDialog;
+import org.rstudio.studio.client.workbench.views.source.model.SourceServerOperations;
 
 public class EditingPreferencesPane extends PreferencesPane
 {
    @Inject
    public EditingPreferencesPane(UIPrefs prefs,
+                                 SourceServerOperations server,
                                  PreferencesDialogResources res)
    {
       prefs_ = prefs;
+      server_ = server;
       PreferencesDialogBaseResources baseRes = PreferencesDialogBaseResources.INSTANCE;
       
       VerticalPanel editingPanel = new VerticalPanel();
@@ -60,8 +70,6 @@ public class EditingPreferencesPane extends PreferencesPane
       editingPanel.add(checkboxPref("Auto-indent code after paste", prefs_.reindentOnPaste()));
       editingPanel.add(checkboxPref("Vertically align arguments in auto-indent", prefs_.verticallyAlignArgumentIndent()));
       editingPanel.add(checkboxPref("Soft-wrap R source files", prefs_.softWrapRFiles()));
-      editingPanel.add(checkboxPref("Ensure that source files end with newline", prefs_.autoAppendNewline()));
-      editingPanel.add(checkboxPref("Strip trailing horizontal whitespace when saving", prefs_.stripTrailingWhitespace()));
       editingPanel.add(checkboxPref(
             "Continue comment when inserting new line",
             prefs_.continueCommentsOnNewline(),
@@ -172,6 +180,61 @@ public class EditingPreferencesPane extends PreferencesPane
             true,
             false);
       displayPanel.add(rmdViewerMode_);
+      
+      
+      VerticalPanel savePanel = new VerticalPanel();
+      
+      savePanel.add(headerLabel("General"));
+      savePanel.add(checkboxPref("Ensure that source files end with newline", prefs_.autoAppendNewline()));
+      savePanel.add(checkboxPref("Strip trailing horizontal whitespace when saving", prefs_.stripTrailingWhitespace()));
+
+      Label serializationLabel = headerLabel("Serialization");
+      serializationLabel.getElement().getStyle().setPaddingTop(14, Unit.PX);
+      savePanel.add(serializationLabel);
+      
+      
+      lineEndings_ = new LineEndingsSelectWidget();
+      spaced(lineEndings_);
+      savePanel.add(lineEndings_);
+      
+      encodingValue_ = prefs_.defaultEncoding().getGlobalValue();
+      savePanel.add(lessSpaced(encoding_ = new TextBoxWithButton(
+            "Default text encoding:",
+            "Change...",
+            new ClickHandler()
+            {
+               public void onClick(ClickEvent event)
+               {
+                  server_.iconvlist(new SimpleRequestCallback<IconvListResult>()
+                  {
+                     @Override
+                     public void onResponseReceived(IconvListResult response)
+                     {
+                        new ChooseEncodingDialog(
+                              response.getCommon(),
+                              response.getAll(),
+                              encodingValue_,
+                              true,
+                              false,
+                              new OperationWithInput<String>()
+                              {
+                                 public void execute(String encoding)
+                                 {
+                                    if (encoding == null)
+                                       return;
+
+                                    setEncoding(encoding);
+                                 }
+                              }).showModal();
+                     }
+                  });
+
+               }
+            })));
+      nudgeRight(encoding_);
+      textBoxWithChooser(encoding_);
+      spaced(encoding_);
+      setEncoding(prefs.defaultEncoding().getGlobalValue());
       
       VerticalPanel completionPanel = new VerticalPanel();
       
@@ -315,6 +378,7 @@ public class EditingPreferencesPane extends PreferencesPane
       tabPanel.setSize("435px", "498px");
       tabPanel.add(editingPanel, "Editing");
       tabPanel.add(displayPanel, "Display");
+      tabPanel.add(savePanel, "Saving");
       tabPanel.add(completionPanel, "Completion");
       tabPanel.add(diagnosticsPanel, "Diagnostics");
       tabPanel.selectTab(0);
@@ -358,6 +422,10 @@ public class EditingPreferencesPane extends PreferencesPane
    @Override
    protected void initialize(RPrefs prefs)
    {
+      // editing prefs
+      EditingPrefs editingPrefs = prefs.getEditingPrefs();
+      lineEndings_.setIntValue(editingPrefs.getLineEndings());
+      
       showCompletions_.setValue(prefs_.codeComplete().getValue());
       showCompletionsOther_.setValue(prefs_.codeCompleteOther().getValue());
       if (prefs_.useVimMode().getValue())
@@ -373,6 +441,11 @@ public class EditingPreferencesPane extends PreferencesPane
    public boolean onApply(RPrefs prefs)
    {
       boolean reload = super.onApply(prefs);
+      
+      // editing prefs
+      prefs.setEditingPrefs(EditingPrefs.create(lineEndings_.getIntValue()));
+                      
+      prefs_.defaultEncoding().setGlobalValue(encodingValue_);
       
       prefs_.codeComplete().setGlobalValue(showCompletions_.getValue());
       prefs_.codeCompleteOther().setGlobalValue(showCompletionsOther_.getValue());
@@ -410,10 +483,21 @@ public class EditingPreferencesPane extends PreferencesPane
    {
       return "Code";
    }
+   
+   private void setEncoding(String encoding)
+   {
+      encodingValue_ = encoding;
+      if (StringUtil.isNullOrEmpty(encoding))
+         encoding_.setText(ChooseEncodingDialog.ASK_LABEL);
+      else
+         encoding_.setText(encoding);
+   }
 
    private final UIPrefs prefs_;
+   private final SourceServerOperations server_;
    private final NumericValueWidget tabWidth_;
    private final NumericValueWidget marginCol_;
+   private final LineEndingsSelectWidget lineEndings_;
    private final NumericValueWidget alwaysCompleteChars_;
    private final NumericValueWidget alwaysCompleteDelayMs_;
    private final NumericValueWidget backgroundDiagnosticsDelayMs_;
@@ -423,6 +507,8 @@ public class EditingPreferencesPane extends PreferencesPane
    private final SelectWidget showCompletionsOther_;
    private final SelectWidget editorMode_;
    private final SelectWidget rmdViewerMode_;
+   private final TextBoxWithButton encoding_;
+   private String encodingValue_;
    
    
 }